sofia-service/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1

psycopg2-binary==2.9.9

google-generativeai==0.8.3
httpx==0.27.2
requests==2.32.3

structlog==24.4.0
